move all store-related things into one folder
