+++ src/app/manage_course_videos/download_all_videos.tsx
@@ -1,9 +1,9 @@
 import React from "react";
 import * as customActions from "../store/custom_actions";
-import * as asyncActions from "../async_actions";
+import * as asyncActions from "../store/async_actions";
 import { useAppDispatch } from "../store/store";
 import { useAppSelector } from "../store/store";
-import { selectCourseVideoStatus } from "../video_selectors";
+import { selectCourseVideoStatus } from "../store/video_selectors";
 import { COURSES_BY_ID } from "../initial_course_list";
 
 interface Props {
