+++ src/app/store/video_download_middleware.ts
@@ -1,5 +1,5 @@
 import { CourseData, VideoQueue } from "../../types";
-import { selectCourseVideoStatus } from "../video_selectors";
+import { selectCourseVideoStatus } from "./video_selectors";
 import VideoDownloader from "../video_downloader";
 import * as customActions from "./custom_actions";
 import { AppMiddleware, RootState } from "./store";
